Add CMakeLists files for sort_pixels_by_bins
