+++ _LowLevelCode/cpp/CMakeLists.txt
@@ -7,6 +7,7 @@ set(
     "hdf_mex_reader"
     "mtimesx_horace"
     "recompute_bin_data"
+    "sort_pixels_by_bins"
 )
 
 foreach(_mex_func ${MEX_FUNCTIONS})
