+++ _LowLevelCode/cpp/sort_pixels_by_bins/CMakeLists.txt
@@ -1,0 +1,16 @@
+set(
+    SRC_FILES
+    "sort_pixels_by_bins.cpp"
+)
+
+set(
+    HDR_FILES
+    "../../build_all/CommonCode.h"
+    "sort_pixels_by_bins.h"
+)
+
+set(MEX_NAME "sort_pixels_by_bins")
+matlab_add_mex(
+    NAME "${MEX_NAME}"
+    SRC "${SRC_FILES}" "${HDR_FILES}"
+)
