--- _LowLevelCode/cpp/sort_pixels_by_bins/sort_pixels_by_bins.sln
@@ -1,26 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual Studio 2008
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "sort_pixels_by_bins", "sort_pixels_by_bins\sort_pixels_by_bins.vcproj", "{2F2ECEEF-3129-4E36-AB1F-005C19F06512}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Debug|Win32.ActiveCfg = Debug|Win32
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Debug|Win32.Build.0 = Debug|Win32
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Debug|x64.ActiveCfg = Debug|x64
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Debug|x64.Build.0 = Debug|x64
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Release|Win32.ActiveCfg = Release|Win32
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Release|Win32.Build.0 = Release|Win32
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Release|x64.ActiveCfg = Release|x64
-		{2F2ECEEF-3129-4E36-AB1F-005C19F06512}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
